+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.5.1] - 2018-01-31
+
+### Added
+
+- Nth item's text step for `ember-power-select`.
+- Error messages now contain detailed information: step, chosen step, args (including label, selector and found elements count). No need to pass those details manually! :D
+
+
+
 ## [0.5.0] - 2018-01-31
 
 ### Added